add tests, cleanup not used classed, lint.
