+++ oauth2_http/javatests/com/google/auth/oauth2/Slf4jUtilsTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024 Google LLC
+ * Copyright 2025 Google LLC
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are
@@ -32,27 +32,16 @@
 package com.google.auth.oauth2;
 
 import static org.junit.Assert.*;
+import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-import ch.qos.logback.classic.spi.ILoggingEvent;
-import com.google.api.client.http.GenericUrl;
-import com.google.api.client.http.HttpRequest;
-import com.google.api.client.http.HttpRequestFactory;
-import com.google.api.client.http.UrlEncodedContent;
-import com.google.api.client.util.GenericData;
-import com.google.gson.JsonParser;
-import com.google.gson.JsonSyntaxException;
-import java.io.IOException;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import com.google.auth.oauth2.Slf4jUtils.LoggerFactoryProvider;
 import org.junit.Before;
 import org.junit.Test;
-import org.mockito.Mockito;
+import org.slf4j.ILoggerFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.slf4j.event.KeyValuePair;
-import org.slf4j.event.Level;
 import org.slf4j.helpers.NOPLogger;
 
 public class Slf4jUtilsTest {
@@ -77,157 +66,35 @@ public class Slf4jUtilsTest {
   }
 
   @Test
-  public void testLogWithMDC_slf4jLogger() {
-
-    TestAppender testAppender = setupTestLogger();
-
-    Map<String, Object> contextMap = new HashMap<>();
-    contextMap.put("key1", "value1");
-    contextMap.put("key2", "value2");
-    Slf4jUtils.logWithMDC(LOGGER, Level.DEBUG, contextMap, "test message");
-
-    assertEquals(1, testAppender.events.size());
-    assertEquals("test message", testAppender.events.get(0).getMessage());
-
-    // Verify MDC content
-    ILoggingEvent event = testAppender.events.get(0);
-    assertEquals(2, event.getMDCPropertyMap().size());
-    assertEquals(ch.qos.logback.classic.Level.DEBUG, event.getLevel());
-    assertEquals("value1", event.getMDCPropertyMap().get("key1"));
-    assertEquals("value2", event.getMDCPropertyMap().get("key2"));
-
-    testAppender.stop();
-  }
-
-  @Test
-  public void testLogWithMDC_INFO() {
-    TestAppender testAppender = setupTestLogger();
-    Slf4jUtils.logWithMDC(LOGGER, Level.INFO, new HashMap<>(), "test message");
-
-    assertEquals(1, testAppender.events.size());
-    assertEquals(ch.qos.logback.classic.Level.INFO, testAppender.events.get(0).getLevel());
-    testAppender.stop();
-  }
-
-  @Test
-  public void testLogWithMDC_TRACE() {
-    TestAppender testAppender = setupTestLogger();
-    Slf4jUtils.logWithMDC(LOGGER, Level.TRACE, new HashMap<>(), "test message");
-
-    assertEquals(0, testAppender.events.size());
-    testAppender.stop();
-  }
-
-  @Test
-  public void testLogWithMDC_WARN() {
-    TestAppender testAppender = setupTestLogger();
-    Slf4jUtils.logWithMDC(LOGGER, Level.WARN, new HashMap<>(), "test message");
-
-    assertEquals(1, testAppender.events.size());
-    assertEquals(ch.qos.logback.classic.Level.WARN, testAppender.events.get(0).getLevel());
-    testAppender.stop();
-  }
-
-  @Test
-  public void testLogWithMDC_ERROR() {
-    TestAppender testAppender = setupTestLogger();
-    Slf4jUtils.logWithMDC(LOGGER, Level.ERROR, new HashMap<>(), "test message");
-
-    assertEquals(1, testAppender.events.size());
-    assertEquals(ch.qos.logback.classic.Level.ERROR, testAppender.events.get(0).getLevel());
-    testAppender.stop();
-  }
-
-  @Test
-  public void testLogGenericData() {
-    // mimic GOOGLE_SDK_JAVA_LOGGING = true
+  public void testGetLogger_loggingEnabled_slf4jBindingPresent() {
     testEnvironmentProvider.setEnv(LoggingUtils.GOOGLE_SDK_JAVA_LOGGING, "true");
     LoggingUtils.setEnvironmentProvider(testEnvironmentProvider);
-
-    TestAppender testAppender = setupTestLogger();
-    GenericData genericData = Mockito.mock(GenericData.class);
-
-    GenericData data = new GenericData();
-    data.put("key1", "value1");
-    data.put("token", "value2");
-
-    LoggerProvider loggerProvider = Mockito.mock(LoggerProvider.class);
-    when(loggerProvider.getLogger()).thenReturn(LOGGER);
-    LoggingUtils.logGenericData(data, loggerProvider, "test generic data");
-
-    assertEquals(1, testAppender.events.size());
-    List<KeyValuePair> keyValuePairs = testAppender.events.get(0).getKeyValuePairs();
-    assertEquals(2, keyValuePairs.size());
-    for (KeyValuePair kvp : keyValuePairs) {
-
-      assertTrue(
-          "Key should be either 'key1' or 'token'",
-          kvp.key.equals("key1") || kvp.key.equals("token"));
-    }
-
-    testAppender.stop();
+    Logger logger = Slf4jUtils.getLogger(LoggingUtilsTest.class);
+    assertNotNull(logger);
+    assertNotEquals(NOPLogger.class, logger.getClass());
   }
 
   @Test
-  public void testLogRequest() throws IOException {
-    // mimic GOOGLE_SDK_JAVA_LOGGING = true
+  public void testGetLogger_loggingEnabled_noBinding() {
     testEnvironmentProvider.setEnv(LoggingUtils.GOOGLE_SDK_JAVA_LOGGING, "true");
     LoggingUtils.setEnvironmentProvider(testEnvironmentProvider);
+    // Create a mock LoggerFactoryProvider
+    LoggerFactoryProvider mockLoggerFactoryProvider = mock(LoggerFactoryProvider.class);
+    ILoggerFactory mockLoggerFactory = mock(ILoggerFactory.class);
+    when(mockLoggerFactoryProvider.getLoggerFactory()).thenReturn(mockLoggerFactory);
+    when(mockLoggerFactory.getLogger(anyString()))
+        .thenReturn(org.slf4j.helpers.NOPLogger.NOP_LOGGER);
 
-    TestAppender testAppender = setupTestLogger();
-    GenericData genericData = Mockito.mock(GenericData.class);
+    // Use the mock LoggerFactoryProvider in getLogger()
+    Logger logger = Slf4jUtils.getLogger(LoggingUtilsTest.class, mockLoggerFactoryProvider);
 
-    GenericData tokenRequest = new GenericData();
-    tokenRequest.set("client_id", "clientId");
-    tokenRequest.set("client_secret", "clientSecret");
-    UrlEncodedContent content = new UrlEncodedContent(tokenRequest);
-
-    MockHttpTransportFactory mockHttpTransportFactory = new MockHttpTransportFactory();
-
-    HttpRequestFactory requestFactory = mockHttpTransportFactory.create().createRequestFactory();
-    HttpRequest request =
-        requestFactory.buildPostRequest(new GenericUrl(OAuth2Utils.TOKEN_SERVER_URI), content);
-
-    LoggerProvider loggerProvider = Mockito.mock(LoggerProvider.class);
-    when(loggerProvider.getLogger()).thenReturn(LOGGER);
-    LoggingUtils.logRequest(request, loggerProvider, "test log request");
-
-    assertEquals(1, testAppender.events.size());
-    assertEquals("test log request", testAppender.events.get(0).getMessage());
-    List<KeyValuePair> keyValuePairs = testAppender.events.get(0).getKeyValuePairs();
-    assertEquals(4, keyValuePairs.size());
-    for (KeyValuePair kvp : testAppender.events.get(0).getKeyValuePairs()) {
-      assertTrue(
-          kvp.key.equals("request.headers")
-              || kvp.key.equals("request.payload")
-              || kvp.key.equals("request.method")
-              || kvp.key.equals("request.url"));
-      if (kvp.key.equals("request.headers") || kvp.key.equals("request.payload")) {
-        assertTrue(isValidJson((String) kvp.value));
-      }
-    }
-    testAppender.stop();
-  }
-
-  boolean isValidJson(String jsonString) {
-    try {
-      JsonParser.parseString(jsonString);
-      return true;
-    } catch (JsonSyntaxException e) {
-      return false;
-    }
+    // Assert that the returned logger is a NOPLogger
+    assertTrue(logger instanceof org.slf4j.helpers.NOPLogger);
   }
 
   @Test
   public void testCheckIfClazzAvailable() {
     assertFalse(Slf4jUtils.checkIfClazzAvailable("fake.class.should.not.be.in.classpath"));
     assertTrue(Slf4jUtils.checkIfClazzAvailable("org.slf4j.event.KeyValuePair"));
-  }
-
-  private TestAppender setupTestLogger() {
-    TestAppender testAppender = new TestAppender();
-    testAppender.start();
-    ((ch.qos.logback.classic.Logger) LOGGER).addAppender(testAppender);
-    return testAppender;
   }
 }
